[svn-r9527] Upgrade hwloc to 1.4.1rc1.

+++ src/pm/hydra/tools/topo/hwloc/hwloc/src/topology-windows.c
@@ -1,6 +1,6 @@
 /*
  * Copyright © 2009 CNRS
- * Copyright © 2009-2011 INRIA.  All rights reserved.
+ * Copyright © 2009-2011 inria.  All rights reserved.
  * Copyright © 2009-2011 Université Bordeaux 1
  * Copyright © 2011 Cisco Systems, Inc.  All rights reserved.
  * See COPYING in top-level directory.
@@ -579,8 +579,8 @@ hwloc_look_windows(struct hwloc_topology *topology)
       }
 
       for (procInfo = procInfoTotal;
-	   (void*) procInfo < (void*) ((unsigned long) procInfoTotal + length);
-	   procInfo = (void*) ((unsigned long) procInfo + procInfo->Size)) {
+	   (void*) procInfo < (void*) ((uintptr_t) procInfoTotal + length);
+	   procInfo = (void*) ((uintptr_t) procInfo + procInfo->Size)) {
         unsigned num, i;
         GROUP_AFFINITY *GroupMask;
